% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_boxplot.R
\name{DRG_boxplot}
\alias{DRG_boxplot}
\title{DRG Boxplot}
\usage{
DRG_boxplot(p1)
}
\arguments{
\item{p1}{choose between 'Average.Covered.Charges', 'Average.Total.Payments', 'Average.Medicare.Payments' to produce a boxplot per each DRG Code. Case-sensitive and include quotations marks}
}
\value{
A boxplot plot of the input column data
}
\description{
This function produces boxplots of the IPPS Dataset for Payments from the 2011 Fiscal Year.
}
\examples{
DRG_boxplot('Average.Covered.Charges')
DRG_boxplot('Average.Total.Payments')
DRG_boxplot('Average.Medicare.Payments')

}
