% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_MedicarePayments_stats.R
\name{DRG_MedicarePayments_stats}
\alias{DRG_MedicarePayments_stats}
\title{Statistics on Total Medicare Payments}
\usage{
DRG_MedicarePayments_stats(p1)
}
\arguments{
\item{p1}{choose between 'Mean', 'Median', 'Standard Deviation' to return the respective statistic for each DRG code. Include quotation marks and case-sensetive.}
}
\value{
A list of called statistics for each DRG code
}
\description{
This function produces statistics about the Total Medicare Payments from IPPS Payments from the Fiscal Year 2011.
}
\examples{
DRG_MedicarePayments_stats('Mean')
DRG_MedicarePayments_stats('Median')
DRG_MedicarePayments_stats('Standard Deviation')

}
